+++ worklet/rollup.config.js
@@ -3,7 +3,7 @@ import typescript from "@rollup/plugin-typescript";
 import terser from "@rollup/plugin-terser";
 
 export default {
-  input: ["DenoiserWorklet.ts"],
+  input: ["RNNoiseWorklet.ts"],
   output: {
     format: "es",
     dir: "../dist",